consumers/requirements.txt
PyYAML==6.0.1
pyspark==3.4.1